Fix issue with Runs Filter dropdown closing whenever filters change  / data refreshes. (#13839)
## Summary & Motivation

The root cause of the issue here is that `<RunsFilterInput />` is being
rendered in two different spots depending on if we have data or if we're
loading. The problem here is that by moving the component to a different
spot in the tree, the component loses all of its state. This PR adds a
hook `usePortalSlot` which allows you to move a component to a different
spot without it losing its state. Based on this tweet:
https://twitter.com/dan_abramov/status/1649464857425485824

## How I Tested These Changes

https://www.loom.com/share/a5601573043246708b8a387865cb82eb