Retry ECS concurrency test on failure
Our use of concurrency here makes this occassionally flaky. I think this
is pretty rare - I've had it running in a tight loop on my local machine
for the past 30 min and I haven't seen it flake yet.

For now, let's just retry on failure using:

https://pypi.org/project/pytest-rerunfailures/

And if it ever re-emerges, we can disable the test or rethink our
approach to detect regressions to:

https://github.com/dagster-io/dagster/pull/11192