[refactor] Delete AssetGroup (#13028)
## Summary & Motivation

Delete `AssetGroup` (in `dagster._legacy`).

From an examination of the code, the only place `AssetGroup` was used in
an internal pathway was during workspace construction. When examining a
Python module, if no job/pipeline/definitions/repo was found, then all
assets would be scraped into a container `AssetGroup`. This just served
to bundle the assets for a short time, until they could be dumped into a
list of repository definitions when converting the loadable target to a
repository.

This PR dispenses with the container `AssetGroup` and just passes a list
containing all of the `AssetsDefinition` and `SourceAsset` objects in
its place. This required only a few adjustments, since none of the
functionality of `AssetGroup` is used anyway.

Aside from these functional changes, this also required updating some
toys and tests to use alternatives to `AssetGroup`, like `Definitions`.

## How I Tested These Changes

Existing test suite.