[ui] Update asset table rows to observe runs and refresh live (#13760)
## Summary & Motivation

Fixes https://github.com/dagster-io/dagster/issues/13749 

Loom video: https://www.loom.com/share/cbd5450e1092412894831046b6fb071a

This PR makes several changes to the asset catalog and location-scoped
asset table:

- Previously the global asset catalog loaded the definitions of all
assets for filtering and search, but the location-scoped asset table
just loaded asset keys. Both passed just the asset key through to the
virtualized row. Both tables now load the definitions at the top level
and pass them in to the rows so they don't have to be re-fetched.

- For SDAs, the per-row query has been replaced with a debounced usage
of `useLiveDataForAssetKeys`, which has some internal fanciness for live
refresh.

- For non-SDAs, the per-row query is still present

To make it possible to put dozens of `useLiveDataForAssetKeys` hooks on
the page at once, I refactored the "run observation" that was part of
that hook into a global. Each hook calls `observeAssetEventsInRuns` to
append itself to a list of subscribers, and a single
`<AssetRunLogObserver />` on the page manages the run log subscriptions.

## How I Tested These Changes

- Updated storybooks

- Visit the global asset catalog and view both SDAs and non-SDAs. Verify
that queries are debounced and that the number of requests in the
network tab does not skyrocket if you scroll quickly through a few pages
of assets.

- Visit the location-scoped asset table

- View a set of assets with a common path root, verify that the "folder"
behavior of the asset catalog looks correct


![image](https://user-images.githubusercontent.com/1037212/233145607-3ec43a99-01c0-465d-b8ea-6cbe355316d1.png)

---------

Co-authored-by: bengotow <bgotow@elementl.com>