[dagit] Fetch freshnessPolicy as part of asset definition, not live data (#13781)
## Summary & Motivation

I noticed this afternoon that we're fetching the `freshnessPolicy` on
assets as part of the "live asset data" instead of loading it along with
other definition level info. This PR just moves the data loading to the
other queries. This means it's not loaded for every asset in the graph
every 15s, which is nice.

## How I Tested These Changes

Verify that storybooks, tests still pass and that you can view the asset
graph, asset catalog, and asset details pages.

---------

Co-authored-by: bengotow <bgotow@elementl.com>