[1/2] Create ScheduledRunListRoot.new  (#13872)
## Summary & Motivation

The "Scheduled" tab on the Runs page has its own root:
`ScheduledRunListRoot`. We need to update this root to look similar to
RunsRootNew.tsx. To prepare for that and to make it easier to review
what actually changed this PR creates a carbon copy of
`ScheduledRunListRoot` which will contain the new changes from the Runs
page redesigns. It's a lot simpler to create a new file than to
conditionally call hooks / add conditional logic everywhere depending on
the feature flag. After we decide to ship these features I'll copy the
contents of the "new" files to the original which will allow us to
preserve commit history.

## How I Tested These Changes

Load dagit successfully.