Display pure source asset cross-repo dependencies (#13828)
Requested by a user:
https://elementl-workspace.slack.com/archives/C0453JAC8KD/p1681818800498509

Currently, we only display cross-repo deps when the asset is defined as
a regular asset (non source asset) in another repository. If it's a
source asset in all repositories, we don't show the cross-repo deps.

This PR amends this behavior to also display cross-repo deps for source
assets that aren't defined as regular assets elsewhere. Also adds this
behavior to the toys for ease of testing.

<img width="800" alt="image"
src="https://user-images.githubusercontent.com/29110579/233715699-c821fdb3-51a5-49fd-abb0-c1fc4e892580.png">