[docs] - Dagster project file reference (#13635)
## Summary & Motivation

This PR adds a reference for files in Dagster projects, including
standard scaffolded files and optional, manually-added files like
`dagster.yaml`. It builds on some of the work Odette did in #13692 and
#13547.

There are some scattered pieces of content that touch on this topic, but
nothing that's a one-stop-shop. The goal of this reference is to explain
the files in a project and provide a jumping-off point for more detailed
content, like how to structure your project or understand the settings
in a configuration file.

Preview:
https://dagster-p5slcysnx-elementl.vercel.app/getting-started/project-file-reference

**Note**: Ideally the `dagster_cloud.yaml` PR would land before this
one, so I can add links in appropriate places. They're not present here
to prevent angering Buildkite tests.

To-do:

- [x] Update the **Create new project** guide
- [x] Update the **Structuring your project** guide

## How I Tested These Changes

:eyes:, local