Guard against empty partition lists
This brings https://github.com/dagster-io/dagster/pull/13844 back but
adds an additional test case for when there are no partitions.

The default implementation already handled this because it first checked
to see if had anything to insert. The new implementation didn't because
without partition keys, it tries to insert an empty list which
SqlAlchemy resolves to trying to insert the default values (a bunch of
nulls).