chore(dbt): constrain dbt-core to support versions 1.2 through 1.5 (#13954)
## Summary & Motivation
Whenever `dbt-core` updates, it inevitably breaks us. Rather than
scrambling to address errors if users did not preemptively pin, or are
unable to conclude from the error messages that they need to pin, we
should just pin on their behalf.

When a user requests support for the next version of dbt Core, that when
we know to loosen this constraint and fix any machinery that needs to be
fixed.

Also, 1.1 is EOL, so bump up the floor to 1.2.

https://docs.getdbt.com/docs/dbt-versions/core

## How I Tested These Changes
bk