Include public members in api docs even if they don't have a docblock
Summary:
User expressed surprise today that the `asset_key` property on DagsterEvent wasn't in the docs - it's a public field, so it seems like it should be in the docs, but it wasn't, because the type wasn't annotated with :undoc-members.

Now that we are filtering all api docs to things with the @public decorater, we might as well include all properties by default, rather than only including the ones with docblocks.

Separately, we should consider requiring that all public properties have docblocks and return types? But indicating their existence without docblocks seems preferable to excluding them.

To do this without confusing sphinx, we had to move some places where we had an attribute defined in the constructor but not in the method to the method instead. (Sphinx was raising warnings about duplicate entries otherwise).

Test Plan: View api docs locally, see many properties that were not there before