enable prefixing source assets when loading (#13920)
## Motivation

From a user:

> Is there a reason why source assets aren't able to be prefixed when
loading from a module?
>
> Context here, is we are trying to load into assets with a client
prefix from the same module per client. We can prepend the client prefix
to normal assets, but it explicitly isn't working for source assets.
Found a comment in the code that seems to indicate this is intentional?

## Summary

This adds a `source_key_prefix` boolean to the core `load_assets_from_*`
functions. This makes it consistent with `load_assets_from_dbt_project`.

## How I Tested These Changes