Only run dependabot on certain folders
Summary:
Running on pyright is leading to a bunch of false positives with transitive dependencies of airflow. Since dependabot doesn't let you exclude folders, enumerate the different top-level folders we do want to run it on.

I also updated the reviewers, although it didn't seem that it was respecting the old setting to make it sandy - hopefully this file is being used at all? Tricky to test without waiting for the next dependabot update.