[docs] - Sidenav improvements, part 3 (#14050)
## Summary & Motivation

This PR builds on #12823 and #13558, making some changes and
improvements to the sidenav:

- Standardizes capitalization (sentence casing) where needed
- Corrects integration name formatting (ex: `Github` > `GitHub`) where
needed
- Moves the **Community** section under **About**
- Updates the tutorial:
- In the sidenav, adds `Part X:` to each part of the tutorial, to make
it clear the progression is sequential
- Re-names the tutorial to `Building your first pipeline using
Software-defined Assets`. While this doesn't get `Pipeline` into the
actual sidenav, it is now in an H1/page title and will be more prominent
in search results for this term.
  
I made this pivot after adding the `Part X:` bits to the sidenav. I
think removing `Tutorial` from the sidenav but retaining these bits
would be an odd experience.
- Adds `Part X:` to the sections for the dbt tutorial to keep things
consistent / clear

## How I Tested These Changes

👀 , local