Refactor job subset specification (#14024)
## Summary & Motivation

The current handling of job subsetting is a bit of a mess due to
inherited technical debt from `PipelineDefinition`. Now that
`PipelineDefinition` is gone, there is a lot of room for improvement.
This PR is the first in a series that will aim to refactor the
subsetting and execution APIs for simplicity, clarity, and consistency.
In this PR:

- Consolidate two subsetting methods `subset_for_execution` and
`subset_for_execution_from_existing_job` on `IJob` class into new
`get_subset` method.
- Replace `IJob.solids_to_execute` with `op_selection`
- Change fields on `ReconstructableJob` to accommodate new interface
- Move op selection resolution to new `op_selection` module. Introduce
`OpSelection` class (currently an implementation detail). This parallels
the existing `AssetSelection` and is part of unifying our approach to
`asset_selection` and `op_selection`.
- `JobDefinition.is_subset_job` -> `is_subset`

## How I Tested These Changes

Existing test suite, with some tests modified to account for changed
APIs.