fix(dagster_airbyte): Add new normalization config check
Fixes #13524.
The Airbyte API changed source of truth for normalization availability in https://github.com/airbytehq/airbyte/pull/21005.
This commit adds an additional check for normalization configuration in
`does_dest_support_normalization`. The current check is kept to
ensure backward compatibility - thus, if any of the checks are true,
the function returns true.