fix(dagster_airbyte): Add new normalization config check (#14174)
## Summary & Motivation
The Airbyte API changed source of truth for normalization availability
in https://github.com/airbytehq/airbyte/pull/21005. This commit adds an
additional check for normalization configuration in
`does_dest_support_normalization`. The current check is kept to ensure
backward compatibility - thus, if any of the checks are true, the
function returns true.

Fixes #13524.

## How I Tested These Changes
I ran the test suite for the `dagster_airbyte` module and added tests
that cover my changes.

Co-authored-by: Nina Jensen <nije@boozt.com>