[dagster-celery] fix inconsistent configuration option (#13800)
## Summary & Motivation

Since commit 3b525b1 the celery executor and corresponding doc have been
updated to the new config structure. Yet the dagster-celery CLI was not
updated to parse the new configuration.

I've updated the CLI to accept the correct configuration, the same one
as the executor. Also, "backward compatibility" is preserved, because
some users may depend on the incorrect implementation.

## How I Tested These Changes

Test cases `engine_config.yaml` and `partial.yaml` are updated, and the
legacy test cases are renamed to `engine_config_legacy.yaml` and
`partial_legacy.yaml`. All tests are green.