Fix yielding run request in sensor result for multipartitioned asset w/ dynamic dimension (#14133)
Reported by a user:
https://dagster.slack.com/archives/C01U954MEER/p1683218607008819

If you have a multipartitioned asset with a dynamic partitions
dimension, yielding a sensor result with a dynamic partitions request
along with a partitioned run request for a multipartitions key with that
new dynamic partition for the dynamic dimension, this currently fails.

This is because we currently validate the set of dynamic partition keys
after requests only when the partitions definition is a dynamic
partitions definition. This PR extends this logic to apply to
multipartitions definitions too.