Make ExecutionPlan build from JobDefinition (#14106)
## Summary & Motivation

`ExecutionPlan` construction does not require access to reconstruction
context, nor does it pass on its job argument to other functions or
structures that need this context. Therefore it makes sense for
construction to operate directly on `JobDefinition`.

This also reroutes all calls to `ExecutionPlan.build` through
`create_execution_plan` (the vast majority of existing calls were
already routed this way but there were a few stragglers). This allows
`ExecutionPlan.build` to be written directly against `JobDefinition`, as
its only callsite is in `create_execution_plan` which should be
considered the "public" function for creating an execution plan.

Prior to this PR, `ExecutionPlan` internal construction operated on an
`IJob`, but the only thing it did was perform a check on
`ReconstructableJob` to make sure its repository load data was the same
as any `repository_load_data` passed in. That's no longer necessary if
we operate on `JobDefinition` directly, and this check is satisfied in
any case by going through `create_execution_plan`.

## How I Tested These Changes

Existing test suite.