[ui] Fix code location expansion key encoding bug (#14217)
## Summary & Motivation

Fix an issue where code location names containing characters that would
be URI-encoded are not expanding correctly in the left nav.

This is because the encoded string for the `RepoAddress` is what we use
as the expansion key, but the code currently checks whether the human
string is expanded.

## How I Tested These Changes

Add a code location that contains a space, "my file.py", alongside the
rest of the toys module. Verify that I can expand/collapse the "my
file.py" code location in the left nav.