Detect places where we are calling capture_error multiple times in a single request
Summary:
Running this middleware twice in a single request is undesirable - sometimes there is other middleware that we hope to run before capture_error in the stack, and if there are multiple capture_errors that becomes impossible.

Detect that case within the middleware and raise an Exception if it happens. I debated logging a warning instead - I can spot-check our capture error calls and make sure that we have test cases for each of the endpoints that have one.

Test Plan: Existing graphql coverage should cover all endpoints that have a capture_error in them today