[ui] Show toast for GraphQL timeouts in Dagit (#11054)
### Summary & Motivation

When the `RootWorkspaceQuery` GraphQL query times out with a 504 gateway
timeout, we currently do nothing to surface this to the user, so it
looks like Dagit is just empty. I'm not sure there's a great way to
surface these errors strictly for this operation, but we could surface a
toast for any 504s to give the user a little bit of context about what's
going on.

To accomplish this, I added a toast for 504/408 errors in the Apollo
error link. I'm not sure this is really the correct thing to do in all
cases, though. It might get pretty noisy in some circumstances.

This change also includes:

- Making the "warning" toast gray instead of red. It's weird that both
"error" and "warning" toasts are currently red.
- Add a little more padding to the toast.
- Redirect to `/locations` in the fallthrough route if there are no code
locations. This is a much better place to dump the user than the weird
mystery empty state we currently show.

### How I Tested These Changes

Return a 504 error for `RootWorkspaceQuery` with a special case in
Python.

Load Dagit, verify that the gray toast appears to indicate the 504.
Verify also that I am redirected to `/locations` from the `/`
fallthrough when no code locations have successfully loaded.