Don't assume run group lists are ordered (#14247)
We recently had a test flake because this comparison assumes the order
of runs in run groups are deterministic. But that doesn't appear to
always be the case:


https://buildkite.com/dagster/internal/builds/43597#01880c16-9d42-4124-8eb2-c39a066c80d0

The alternative is to force run groups to return in a deterministic
order. I don't know enough about how they're used to know if that's the
behavior we want.