[ui] Show an instructional tooltip about zoom modifier keys (#14314)
## Summary & Motivation

This is a small tooltip that appears beside the zoom control when you
start using the scroll wheel (to attempt to zoom) and disappears for the
remainder of the time the component is mounted once you use a modifier
key. Our goal here is to compensate for the lack of discoverability of
the zoom functionality for users who do not have trackpads (pinch to
zoom) and are not familiar with Figma, Photoshop, et. al.

https://www.loom.com/share/e206ec32d0724607ac45703ef0fc1d12


## How I Tested These Changes

I verified this manually in Chrome, Firefox, Edge and Safari. I used
user-agent overrides in Chrome to verify that the key changes from "Cmd"
to "Ctrl" for non-Mac users.

---------

Co-authored-by: bengotow <bgotow@elementl.com>