Do not initialize instance in cases where schedules and sensors have no resources defined (#14335)
## Summary & Motivation

A user reported that during an upgrade from 1.1.9 to 1.3.4 they were
required to access an environment variable they had not been required to
access before. The issue is the change to support resources in schedules
and sensors added a code path that unconditionally constructed the
instance where it did not before. This PR reverts to the old behavior.

However I do not think this is the right fix. Allowing users to deploy
code locations that do not have access to the instance is extremely
fragile. Any access to `context.instance` can trigger a fatal, and if we
add any feature at any time.

The core problem here is the wildly complicated state of all our
initialization pathways, where everything is "Optional" when in fact
they are not. It seems that these evolved this way to enable convenient
test pathways, or just out of sheer laziness, but we are now left in a
state where there is a combinatorial explosion of possible states in all
of our context objects, and further more on top of that a combinatorial
explosion of initialization codepaths, as we seem to lazily create
things whenever it pleases us. This will result in a continual stream of
difficult-to-predict bugs as it is super fragile, as this bug
demonstrates.

## How I Tested These Changes

Included unit tests