Actually apply runtime config to instantiated ConfigurableResource objects (#14306)
Fix for
https://github.com/dagster-io/dagster/issues/14309#event-9257838645.

This PR untangles and fixes a few separate bugs, which were all
obfuscated by the fact that were calling `process_config` on the config
passed to a `ConfigurableResource` at instantiation:
- We were previously ignoring unstructured run config passed into a
configurable resource. This PR fixes to dump config values into said
configurable resource.
- We were erroneously allowing `EnvVar` and `IntEnvVar` to be passed in
unstructured run config - and then promptly treating it as just a
string. This PR adds an error message to point users in correct
directions.
- Enums were being incorrectly converted from pythonic space to legacy
config space - meaning you would receive different config values
depending on where an enum value was set. This PR unifies to always use
`dagster.Enum.from_python_enum` to perform pythonic enum to dagster enum
conversion. Furthermore, it changes `from_python_enum` to set the
dagster enum's values to be the values of the pythonic enum, rather than
the _symbolic_ values. This avoids the weird multiple enums in same
python execution bug I discuss later.
- The resolution of env vars in `_resolve_and_update_env_vars` was
actually the only way that nested resource config was ever getting
post-processed - this explicitly adds a post-processing step for that
configuration (at resource initialiation time)

I also uncovered something wacky going on with having multiple enums of
the same name in the same file - it seems to cause pydantic models to
not be able to track equality between those enums after config
translation, and causes tests to fail with a pydantic validation error.
Because of the fact that we were passing only the enum _values_ and not
the symbolic value, we weren't hitting this weird equality thing.
Instead forcing everything use enum values only, means we never hit
this.

---------

Co-authored-by: Nicholas Schrock <schrockn@elementl.com>