[ui] Fix Run timeline resizing (#14339)
## Summary & Motivation

When I refactored the run timeline to be virtualized, I broke the
resizing behavior that expands/shrinks the rows to fit the window.
Restore this. ~and apply a debounce to the width setter so that the
resize doesn't risk a crunchy animation.~

## How I Tested These Changes

View run timeline in app and in Storybook, resize window. Verify that
the timeline rows and chunks re-render properly.