RFC: Add .env file env vars in all our CLI entrypoints if there's a .env file in the local directory
Summary:
This PR intends to address a common point of user confusion in support - we tell users in our docs to put env vars in a .env file and they'll be automatically injected into the environment, but it doesn't apply to our own dagster env vars, just the ones that are used in their code. So users still hit the same friction / confusion that the .env file is intended to address.

So instead of using a pluggable component and defaulting it to a SecretLoader that pulls from a local .env file but is only called in certain CLIs, make it called in every CLI / Dagster entrypoint.

I thought about putting this directly in DagsterInstance.get() but it seemed bad to add a permanent side effect there that wasn't there before, and there are many places where we create a DagsterInstance outside of a process entrypoint that wouldn't need to inject the env vars.

Will add testing after any initial thoughts on this direction.