feat(dbt-assets)[1/N]: add dbt_asset decorator (#14186)
## Summary & Motivation
Begin the scaffold for a `dbt_asset` decorator. The decorator takes in a
dbt manifest and dbt select statement syntax and creates the
corresponding asset definitions for the dbt models. In later diffs in
the stack, we provide support for the user to populate the compute
function for the decorator.

## How I Tested These Changes
pytest