Fix subsetting bug (#14390)
## Summary & Motivation

We have some special handling in place for the rare case that you have a
subsettable multi asset which has some nodes upstream of another asset
(`A`) and other nodes downstream of `A`.

In these cases, we need to separate out the asset into two separate op
invocations in order for it to be possible to execute (as otherwise the
op for the multi-asset would need to execute both before and after the
op for `A`).

In some cases, this requires just invoking the original op two separate
times, but in other cases, we actually need to generate a new op. See:
https://github.com/dagster-io/dagster/pull/13830 for more context.

However, we were overly zealous in the original PR, generating new ops
in situations where it was not actually necessary (as the broader job
would not actually need to plug in any output into the new input we
created).

This PR fixes this, meaning we will only create the new input in cases
where it will actually be used.

Further work should be done to investigate if there are other similar
issues that are caused by this behavior, but this PR should quarantine
the impact to the rare situation described in the above PR.

## How I Tested These Changes

Modified test failed before this change, passes now.