Elevate pin to dagster
The primary motivation here is to fix our tests because it looks like
upstream projects are working on their own fixes.

If Thursday's release comes and a fix still isn't available, we'll ship
with this pin. But I'd prefer if we could remove this before then.

In the meantime, users who encounter the issue can fix by pinning
typing_expectations < 4.6.0.