Prefer run config over config mapping when reconstructing Jobs (#14392)
## Summary

When a job is supplied a run config dict, we use
`_config_mapping_with_default_value` to create a corresponding config
mapping which produces this default config. This config mapping also
bundles in information related to e.g. the executor.
The job object then has `_run_config` and `_config_mapping` set.

When copying a job (to attach a default executor, for example), we
forward the config by passing `job._config_mapping or job._run_config`.

In the case where the original job object was passed a run config dict,
it has `_run_config` and `_config_mapping` set. The or precedence means
the copied job has `_config_mapping`, which does not take into account
any changes in the copy (such as a new executor).

This PR reverses the `or` precedence so that `_run_config` is passed and
a new config mapping is generated in this scenario.

Should resolve #14371

## Test Plan

Reproduced #14371 locally and ensured this fixed. Added a small test
case (previously failing).