[ui] Repair path-prefix (#14421)
## Summary & Motivation

With the upgrade to Storybook, `@types/webpack` was eliminated, which
removed the global type declaration for `__webpack_public_path__`. In an
effort to repair the resulting type error, I added a `(window as any)`
to the variable in `publicPath.ts`. It appears that this is what led to
the breakage of path prefixing.

By manually adding a type declaration for this Webpack global in our own
typedef file, I can remove the `(window as any)` annotation, and the
variable goes back to working properly.

## How I Tested These Changes

`yarn build`, load dagit with a path-prefix. Verify that the app loads,
and that all JS assets are retrieved properly.