feat(dbt): automatically apply asset key replacements using the manifest (#14428)
## Summary & Motivation
Because we have access to the manifest, the source of truth of mappings,
and the underlying `AssetsDefinition`, just apply the new mappings onto
the created definition. We do not need to wait for the user to do it.

Originally, this was going to live in the `.with_attributes` decorator,
but this seems more ergonomic.

## How I Tested These Changes
pytest