[Cancel asset backfill 1/2] Enable cancel backfill button (#14244)
This PR adds support for canceling an asset backfill:
- Updates the `numCancelable` resolver to return the number of
partitions for which runs have not yet been created. This is used to
display a "There are X partitions yet to be queued or launched"
confirmation message.
- Validates permissions and updates the backfill's bulk action status to
be `CANCELED`.

One gap following this PR is that after the asset backfill is marked as
canceled the backfill is not evaluated by the daemon again. This means
that asset backfill page will continue to display the state of the
backfill at the moment of cancellation, even if the runs have since
finished. Part 3 of this stack resolves this gap:
https://github.com/dagster-io/dagster/pull/14224

<img width="1225" alt="image"
src="https://github.com/dagster-io/dagster/assets/29110579/51fde917-c08f-4516-9762-63cf56e40fa6">