[Cancel asset backfill 2/2] Cancel runs within daemon and update asset backfill data until complete (#14224)
Following https://github.com/dagster-io/dagster/pull/14244:
- Unfinished runs continue to execute
- The asset backfill page does not update after a backfill is canceled
by a user. This is because the daemon stops evaluating backfills with a
canceled status, so the backfill page continues to display the status of
the materialized/failed asset partitions at the moment of cancellation.

This PR bridges the two gaps above by introducing a new `CANCELING` bulk
action status. This status represents asset backfills that have been
requested for cancellation by a user, but whose runs have not yet
successfully canceled. In this `CANCELING` state, the daemon:
- Does not create or queue new runs
- Marks existing runs as canceled when possible
- Updates asset backfill data with materialized/failed partitions until
all runs have concluded
- Marks the backfill is as `CANCELED` when all requested runs have
concluded. This allows the asset backfill page to continue to update
until all runs have finished.

The reason why the daemon is responsible for canceling the runs is
because runs with certain statuses cannot be canceled; only runs in the
`STARTED` or `QUEUED` state can be canceled. The other unfinished states
(`STARTING`, `NOT_STARTED`) will eventually move to a `STARTED` or
`QUEUED` state, allowing them to be canceled. Handling this within the
daemon allows us to reevaluate un-cancellable runs at a later time and
cancel them when possible.