[ui] Guard against partition set status error (#14479)
## Summary & Motivation

An error can occur on the op job partitions view when the partition
status field returns with an error. We have no way to map the partitions
back to the statuses, and an unreachable assert is hit because we try to
evaulate a status string on `undefined`.

Work around this by allowing `undefined` to be treated as a "status",
rendering as dark gray and using the label "Unknown". This should
ideally not happen often, and seems better than erroring out the whole
page because of the faulty error handling.

## How I Tested These Changes

Added a jest spec with a fragment matching the error state.

Raise an error from `resolve_partitionStatusesOrError` in Python and
load an op job partitions view. Verify that the partition bar loads
entirely dark gray, but that the page doesn't crash.

<img width="1296" alt="Screenshot 2023-05-25 at 8 58 45 PM"
src="https://github.com/dagster-io/dagster/assets/2823852/9ab8593d-ad02-420a-907d-10e5832dd561">